Restructure library code to create LDAP user
