+++ Pipfile
@@ -6,7 +6,7 @@ name = "pypi"
 [packages]
 django-admin-list-filter-dropdown = "==1.0.3"
 django-admin-rangefilter = "==0.6.2"
-django-auth-ldap = "==2.2.0"
+django-auth-ldap = "==4.3.0"
 django-bootstrap4 = "==2.2.0"
 django-export-action = "==0.1.1"
 django-hashers-passlib = "*"
