+++ humans/templates/shipanaro/index.html
@@ -23,9 +23,25 @@
       </div>
     </div>
   </div>
-</div>
-<div class="row">
   <div class="col-lg mb-4">
+    <div class="card">
+      <div class="card-body">
+        <h5 class="card-title">{% trans 'Cloud' %}</h5>
+        <p class="card-text">{% trans "Repositori de fitxers i documents." %}</p>
+        <a href="https://nuvol.pirates.cat/" class="btn btn-primary">{% trans 'Go' %} <i class="fas fa-angle-double-right"></i></a>
+      </div>
+    </div>
+  </div>
+  <div class="col-lg mb-4">
+    <div class="card">
+      <div class="card-body">
+        <h5 class="card-title">{% trans 'Xat' %}</h5>
+        <p class="card-text">{% trans "Eina de missatgeria instantània per a grups." %}</p>
+        <a href="https://xat.pirates.cat" class="btn btn-primary">{% trans 'Go' %} <i class="fas fa-angle-double-right"></i></a>
+      </div>
+    </div>
+  </div>
+  {% comment %} <div class="col-lg mb-4">
     <div class="card">
       <div class="card-body">
         <h5 class="card-title">{% trans 'Sistema de Participació' %}</h5>
@@ -44,8 +60,6 @@
       </div>
     </div>
   </div>
-</div>
-<div class="row">
   <div class="col-lg mb-4">
     <div class="card">
       <div class="card-body">
@@ -60,31 +74,11 @@
   <div class="col-lg mb-4">
     <div class="card">
       <div class="card-body">
-        <h5 class="card-title">{% trans 'Xat' %}</h5>
-        <p class="card-text">{% trans "Eina de missatgeria instantània per a grups." %}</p>
-        <a href="https://xat.pirates.cat" class="btn btn-primary">{% trans 'Go' %} <i class="fas fa-angle-double-right"></i></a>
-      </div>
-    </div>
-  </div>
-</div>
-<div class="row">
-  <div class="col-lg mb-4">
-    <div class="card">
-      <div class="card-body">
         <h5 class="card-title">{% trans 'Wiki' %}</h5>
         <p class="card-text">{% trans "Wiki del partit." %}</p>
         <a href="https://wiki.pirates.cat/" class="btn btn-primary">{% trans 'Go' %} <i class="fas fa-angle-double-right"></i></a>
       </div>
     </div>
-  </div>
-  <div class="col-lg mb-4">
-    <div class="card">
-      <div class="card-body">
-        <h5 class="card-title">{% trans 'Cloud' %}</h5>
-        <p class="card-text">{% trans "Repositori de fitxers i documents." %}</p>
-        <a href="https://nuvol.pirates.cat/" class="btn btn-primary">{% trans 'Go' %} <i class="fas fa-angle-double-right"></i></a>
-      </div>
-    </div>
-  </div>
+  </div> {% endcomment %}
 </div>
 {% endblock %}
